Circular Regression

To be used similar to the stats::lm() function. This function assumes that the dependent variable has a circular pattern (e.g. time, angles, directions). The independent variable may either all be circular, or they may all be linear (but not both).

cm(
  formula,
  type,
  data,
  units,
  tidy = FALSE,
  conf.int = TRUE,
  conf.level = 0.95,
  ...
)

Arguments

formula

Standard formula with single dependent variable and multiple independent variables

type

describes the right-hand side of the formula as being either "linear" or "circular" distributions

data

Data set that contains the named terms from the formula

units

The type of units that the circular data should be calculated in (e.g. radians, degrees)

tidy

If the results should be in a text block or should be in a tidy table. If FALSE, then the confidence arguments are not needed/used.

Details

Other standards for these types of models have been preset to simplify modeling. This function requires installation of the {circular} package for the back-end of this function.

Examples

formula = el_svg ~ lab_hba1c + bmi
type = "linear"
data = card::geh
#> Error in loadNamespace(x): there is no package called ‘card’
units = "degrees"
raw <- cm(formula = formula, type = type, data = data, units = units)
#> Error in loadNamespace(x): there is no package called ‘circular’
tidy <- cm(formula = formula, type = type, data = data, units = units, tidy = TRUE)
#> Error in loadNamespace(x): there is no package called ‘circular’