Visualilzation of EGM using ggplot

Visualilzation of EGM using ggplot

ggm(
  data,
  channels,
  time_frame = NULL,
  annotation_channel = NULL,
  beats = NULL,
  intervals = FALSE,
  hline = FALSE,
  vline = FALSE,
  ...
)

Arguments

data

Long-format data for visualization. It is a combined format that includes both signal data and header information. Can be generated by the melt_egm() function.

channels

Character vector of which channels to use. Can give either the channel label (e.g "CS 1-2") or the recording device/catheter type (e.g "His" or "ECG").

time_frame

A time range that should be displaced given in the format of a vector with a length of 2. The left value is the start, and right value is the end time. This is given in seconds (decimals may be used).

annotation_channel

Identifies which of the selected channels will be used for annotation. This channel will be used for the analysis of intervals, peaks, etc. If NULL, no annotations will be made.

beats

A integer vector that represents the indexed beats that should be analyzed within the plot. Each beat is indexed as 1...n from what is within the time_frame indicated. For example, if 5 beats are seen, beats = c(1, 5) would select the 1st and 5th beat. This can also refer to any large deflection that is consistent an EGM.

intervals

The choice of whether interval data will be included. An annotation channel must be present, otherwise nothing will be plotted. This argument allows several choices.

FALSE: no intervals will be annotated, and is the default option

TRUE: all intervals will be annotated

integer: an integer vector that represents the indexed intervals that should be annotated. If NULL, no intervals will be annotated. For example, if their are 5 beats, than there are 4 intervals between them that can be labeled. They can be referenced by their index, e.g. intervals = c(2) to reference the 2nd interval in a 5 beat range.